requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Naked==0.1.31
PyYAML==5.1.1
requests==2.22.0
SQLAlchemy==1.3.5
urllib3==1.25.3
waitress==1.3.0
Werkzeug==0.15.4
